+++ src/app/(afterlogin)/presentation/upload/[id]/_component/InputSection.module.scss
@@ -60,7 +60,3 @@
     background-color: #d9d9d9;
   }
 }
-
-.modalContent {
-  margin: 0;
-}
