--- src/app/_components/_modules/AlertModal.module.scss
@@ -1,50 +1,0 @@
-@import '@/app/_components/globals';
-
-@keyframes fade-in {
-  from {
-    opacity: 0;
-  }
-
-  to {
-    opacity: 1;
-  }
-}
-
-@keyframes fade-out {
-  from {
-    opacity: 1;
-  }
-
-  to {
-    opacity: 0;
-  }
-}
-
-.modalContainer {
-  position: absolute;
-  top: 40%;
-  left: 50%;
-  width: 250px;
-  height: 35px;
-  transform: translate(-50%, -50%);
-  animation: fade-in 0.5s;
-}
-
-.modalContainer.fadeOut {
-  animation: fade-out 0.8s; // 0.5보다 조금 더 늘려야 함
-}
-
-.modalContent {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 20px;
-  width: 100%;
-  height: 100%;
-  background-color: darkgray;
-  border-radius: 20px;
-
-  .closeButton {
-    @include pure-button;
-  }
-}
