--- src/app/_components/_modules/ConfirmModal.module.scss
@@ -1,21 +1,0 @@
-.modalContainer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: absolute;
-  inset: 0;
-  background-color: rgb(0 0 0 / 50%);
-}
-
-.modalContent {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-  top: 5%;
-  width: 500px;
-  height: 700px;
-  background: #fff;
-  border-radius: 20px;
-  flex-direction: column;
-}
