+++ src/app/_components/_modules/_modal/AlarmToast.module.scss
@@ -1,0 +1,19 @@
+.modalContent {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  position: fixed;
+  top: 15%;
+  left: 50%;
+  width: 250px;
+  height: 35px;
+  transform: translate(-50%, -50%);
+  gap: 20px;
+  background-color: darkgray;
+  border-radius: 20px;
+
+  p {
+    margin: 0;
+    color: white;
+  }
+}
