+++ src/app/_components/_modules/_modal/AlarmToast.tsx
@@ -1,0 +1,12 @@
+'use client';
+import styles from './AlarmToast.module.scss';
+
+const AlarmToast = () => {
+  return (
+    <div className={styles.modalContent}>
+      <p>알람</p>
+    </div>
+  );
+};
+
+export default AlarmToast;
