+++ src/app/_components/_modules/_modal/SaveToast.tsx
@@ -1,0 +1,12 @@
+'use client';
+import styles from './SaveToast.module.scss';
+
+const SaveToast = () => {
+  return (
+    <div className={styles.modalContent}>
+      <p>저장이 완료되었습니다</p>
+    </div>
+  );
+};
+
+export default SaveToast;
