+++ src/app/_components/_modules/_modal/Toast.module.scss
@@ -1,0 +1,33 @@
+@import '@/app/_components/globals';
+
+@keyframes fade-in {
+  from {
+    opacity: 0;
+  }
+
+  to {
+    opacity: 1;
+  }
+}
+
+@keyframes fade-out {
+  from {
+    opacity: 1;
+  }
+
+  to {
+    opacity: 0;
+  }
+}
+
+.toastContainer {
+  // position: fixed;
+  // top: 15%;
+  // left: 50%;
+  // transform: translate(-50%, -50%);
+  animation: fade-in 0.5s;
+
+  &.fadeOut {
+    animation: fade-out 0.8s;
+  }
+}
